// src/app/aura/hr/enum/holidays.enum.ts
export enum Holidays {
    NewYearDay = 'New Year Day',
    MemorialDay = 'Memorial Day',
    IndependenceDay = 'Independence Day',
    LaborDay = 'Labor Day',
    Diwali = 'Diwali',
    ThanksgivingDay = 'Thanksgiving Day',
    BlackFriday = 'Black Friday',
    ChristmasEve = 'Christmas Holiday',
    MartinLutherKing = 'Martin Luther King Day',
}